+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -343,7 +343,10 @@ requireGW(
           model.hoverCard(undefined);
         }, 300);
         return;
-      } else {
+      } else if (card.id() === "gwc_minion")
+        model.gwaioShowTechCardTooltip()[3] = false;
+      else {
+        model.gwaioShowTechCardTooltip()[3] = true;
         makeCardTooltip(card);
       }