+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unrealted
+
+### Fixed
+
+- Hide tooltip from Sub Commander cards in your inventory
+
 ## v4.7.1 - 2020-08-24
 
 ### Fixed